+++ priv/repo/migrations/20260308164703_add_inserted_at_index_to_archives.exs
@@ -1,0 +1,10 @@
+defmodule Linkhut.Repo.Migrations.AddInsertedAtIndexToArchives do
+  use Ecto.Migration
+
+  @disable_ddl_transaction true
+  @disable_migration_lock true
+
+  def change do
+    create index(:archives, [:inserted_at], concurrently: true)
+  end
+end